pyproject.toml
[tool.poetry]
name = "llmpeg"
version = "0.1.0"
description = ""
authors = ["rodfer0x80 <rodrigolf080@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11.0, <3.12" 
tts = "^0.22.0"
selenium = "^4.21.0"
jsonargparse = "^4.28.0"
ollama = "^0.2.0"
openai-whisper = "^20231117"
sounddevice = "^0.4.6"
soundfile = "^0.12.1"
numpy = "^1.26.4"
python-vlc = "^3.0.20123"
beautifulsoup4 = "^4.12.3"
yt-dlp = "^2024.4.9"


easyocr = "^1.7.1"
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
main = "llmpeg.__main__:main"
